attempt move of files to correct places
